Html in code and new handler
