--- server/share.html
@@ -1,20 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <meta http-equiv="refresh" content="0;url={{.RedirectURL}}">
-
-    <meta property="og:locale" content="pt_BR">
-    <meta property="og:site_name" content="Vitrine Social">
-    <meta property="og:type" content="website">
-
-    <meta property="og:url" content="{{.MetaURL}}">
-    <meta property="og:title" content="{{.MetaTitle}}">
-
-    <meta property="og:description" content="{{.MetaDescription}}">
-
-    <meta property="og:image" content="{{.MetaImage}}">
-    <meta property="og:image:type" content="image/jpeg">
-  </head>
-  <body>
-  </body>
-</html>
